Move player class from tictactoe to tictactoehandle
